Figured out the favicon
I just wanted to change the favicon.

I have done a lot of good work. It is looking really good now.

Still lots to do. What would be really cool is to create a python script or something to scrape a GitHub page and create a new post for me. But that is something I could do for sure.